+++ messages/ar/base.php
@@ -111,4 +111,10 @@ return [
     'You cannot move the folder "{name}"!' => 'لا يمكنك نقل المجلد "{name}"!',
     'ZIP selected' => 'تم تحديد ملف مضغوط ZIP',
     'ZIP support is not enabled.' => 'دعم الملفات المضغوطة zip غير مفعل.',
+    'Could not import file with guid {guid}. File not found' => '',
+    'File (Module)' => '',
+    'File content ID' => '',
+    'Files (Module)' => '',
+    'Folder content ID' => '',
+    'Folders' => '',
 ];