+++ messages/de/base.php
@@ -111,4 +111,10 @@ return [
     'You cannot move the folder "{name}"!' => 'Du kannst den Ordner "{name} nicht verschieben"!',
     'ZIP selected' => 'Ausgewählte komprimieren (ZIP)',
     'ZIP support is not enabled.' => 'ZIP-Unterstützung ist nicht aktiviert.',
+    'Could not import file with guid {guid}. File not found' => '',
+    'File (Module)' => '',
+    'File content ID' => '',
+    'Files (Module)' => '',
+    'Folder content ID' => '',
+    'Folders' => '',
 ];