+++ messages/es/base.php
@@ -111,4 +111,10 @@ return [
     'You cannot move the folder "{name}"!' => '¡No puedes mover la carpeta "{name}"!',
     'ZIP selected' => 'Comprimir archivos seleccionados',
     'ZIP support is not enabled.' => 'Soporte ZIP no está activado.',
+    'Could not import file with guid {guid}. File not found' => '',
+    'File (Module)' => '',
+    'File content ID' => '',
+    'Files (Module)' => '',
+    'Folder content ID' => '',
+    'Folders' => '',
 ];